+++ src/storage/store/memory.rs
@@ -7,9 +7,8 @@ use crate::{
     array::MaybeBytes,
     byte_range::{ByteOffset, ByteRange},
     storage::{
-        ListableStorageTraits, ReadableListableStorageTraits, ReadableStorageTraits,
-        ReadableWritableStorageTraits, StorageError, StoreKeyRange, StoreKeyStartValue,
-        StoreKeysPrefixes, WritableStorageTraits,
+        ListableStorageTraits, ReadableStorageTraits, StorageError, StoreKeyRange,
+        StoreKeyStartValue, StoreKeysPrefixes, WritableStorageTraits,
     },
 };
 
@@ -209,10 +208,6 @@ impl ListableStorageTraits for MemoryStore {
         Ok(StoreKeysPrefixes { keys, prefixes })
     }
 }
-
-impl ReadableWritableStorageTraits for MemoryStore {}
-
-impl ReadableListableStorageTraits for MemoryStore {}
 
 #[cfg(test)]
 mod tests {