+++ CHANGELOG.md
@@ -34,6 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Added a fast path to `Array::retrieve_array_subset` methods if the array subset matches a chunk
  - Added `ZARR_NAN_F64/F32/F16/BF16` which match the zarr nan bit representation on all implementations
  - Added `size_usize()` to `ArrayRepresentation`
+ - Add generic implementations of storage supertraits (`ReadableWritableStorageTraits` and `ReadableListableStorageTraits`)
 
 ### Changed
  - **Breaking**: `array::data_type::DataType` is now marked `#[non_exhaustive]`